tests/wpt/web-platform-tests/css/css-flexbox/getcomputedstyle/flexbox_computedstyle_display-inline.html

<!DOCTYPE html>
<title>flexbox | computed style | display: inline-flex</title>
<link rel="author" href="http://opera.com" title="Opera Software">
<link rel="help" href="http://www.w3.org/TR/css-flexbox-1/#flex-containers">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name="flags" content="dom">
<style>
body {
	color: red;
	display: inline-flex;
}
</style>
<div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("display"),
	          "inline-flex");
});
</script>
